# docs/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/providers": {
      "get": {
        "operationId": "ProvidersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ProvidersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}": {
      "get": {
        "operationId": "ProvidersController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProvidersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}/status": {
      "patch": {
        "operationId": "ProvidersController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProviderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}/capabilities": {
      "post": {
        "operationId": "ProvidersController_declareCapability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclareProviderCapabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}/integrations": {
      "post": {
        "operationId": "ProvidersController_createIntegration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderIntegrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}/licenses": {
      "post": {
        "operationId": "ProvidersController_createLicense",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderLicenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/providers/{id}/attribution-rules": {
      "post": {
        "operationId": "ProvidersController_createAttributionRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderAttributionRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/admin/provider-licenses/{id}": {
      "get": {
        "operationId": "ProviderLicensesController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProviderLicensesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderLicenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/provider-licenses/{id}/rights": {
      "patch": {
        "operationId": "ProviderLicensesController_setRights",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProviderLicenseRightsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/provider-licenses/{id}/status": {
      "patch": {
        "operationId": "ProviderLicensesController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProviderLicenseStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/provider-licenses/{id}/evidence": {
      "post": {
        "operationId": "ProviderLicensesController_addEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderPolicyEvidenceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/admin/provider-licenses/{id}/data-policy": {
      "post": {
        "operationId": "ProviderLicensesController_createDataPolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDataPolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-providers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Dau Viet API",
    "description": "Living Digital Atlas of Vietnam - backend API contract for web, mobile, and admin clients. See docs/backend/BACKEND_HANDOFF.md for the full handoff contract this document is generated alongside.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "GOOGLE_PLACES",
            "description": "Stable, machine-readable, immutable after creation."
          },
          "name": {
            "type": "string"
          },
          "websiteUrl": {
            "type": "string"
          },
          "developerUrl": {
            "type": "string"
          },
          "partnerPortalUrl": {
            "type": "string"
          },
          "credentialMode": {
            "type": "string",
            "enum": [
              "NONE",
              "API_KEY",
              "OAUTH2",
              "PARTNER_CONTRACT",
              "OTHER"
            ]
          },
          "supportedEnvironments": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SANDBOX",
                "PRODUCTION"
              ]
            }
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "websiteUrl": {
            "type": "string"
          },
          "developerUrl": {
            "type": "string"
          },
          "partnerPortalUrl": {
            "type": "string"
          },
          "credentialMode": {
            "type": "string",
            "enum": [
              "NONE",
              "API_KEY",
              "OAUTH2",
              "PARTNER_CONTRACT",
              "OTHER"
            ]
          },
          "supportedEnvironments": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SANDBOX",
                "PRODUCTION"
              ]
            }
          }
        }
      },
      "SetProviderStatusDto": {
        "type": "object",
        "properties": {}
      },
      "DeclareProviderCapabilityDto": {
        "type": "object",
        "properties": {
          "capability": {
            "type": "string",
            "enum": [
              "PLACE_SEARCH",
              "PLACE_DETAIL",
              "NEARBY_SEARCH",
              "TEXT_SEARCH",
              "ACCOMMODATION_SEARCH",
              "ACCOMMODATION_DETAIL",
              "LIVE_PRICE",
              "AVAILABILITY",
              "RESTAURANT_SEARCH",
              "RESTAURANT_DETAIL",
              "ACTIVITY_SEARCH",
              "ACTIVITY_DETAIL",
              "FLIGHT_SEARCH",
              "CAR_RENTAL_SEARCH",
              "REVIEWS",
              "PHOTOS",
              "BOOKING_REDIRECT",
              "BOOKING_API",
              "AFFILIATE_LINK",
              "WEBHOOK",
              "CONVERSION_REPORTING"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "capability"
        ]
      },
      "CreateProviderIntegrationDto": {
        "type": "object",
        "properties": {
          "environment": {
            "type": "string",
            "enum": [
              "SANDBOX",
              "PRODUCTION"
            ]
          },
          "credentialReference": {
            "type": "string",
            "description": "An env-var / secret-manager KEY NAME only - never the secret value itself."
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "environment"
        ]
      },
      "CreateProviderLicenseDto": {
        "type": "object",
        "properties": {
          "datasetOrProduct": {
            "type": "string",
            "example": "Places API core data"
          },
          "capability": {
            "type": "string",
            "enum": [
              "PLACE_SEARCH",
              "PLACE_DETAIL",
              "NEARBY_SEARCH",
              "TEXT_SEARCH",
              "ACCOMMODATION_SEARCH",
              "ACCOMMODATION_DETAIL",
              "LIVE_PRICE",
              "AVAILABILITY",
              "RESTAURANT_SEARCH",
              "RESTAURANT_DETAIL",
              "ACTIVITY_SEARCH",
              "ACTIVITY_DETAIL",
              "FLIGHT_SEARCH",
              "CAR_RENTAL_SEARCH",
              "REVIEWS",
              "PHOTOS",
              "BOOKING_REDIRECT",
              "BOOKING_API",
              "AFFILIATE_LINK",
              "WEBHOOK",
              "CONVERSION_REPORTING"
            ],
            "description": "Leave unset for a dataset-wide license not tied to one capability."
          },
          "licenseType": {
            "type": "string"
          },
          "termsUrl": {
            "type": "string"
          },
          "privacyUrl": {
            "type": "string"
          },
          "developerTermsUrl": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string"
          },
          "effectiveUntil": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "datasetOrProduct",
          "termsUrl"
        ]
      },
      "CreateProviderAttributionRuleDto": {
        "type": "object",
        "properties": {
          "licenseId": {
            "type": "string"
          },
          "capability": {
            "type": "string",
            "enum": [
              "PLACE_SEARCH",
              "PLACE_DETAIL",
              "NEARBY_SEARCH",
              "TEXT_SEARCH",
              "ACCOMMODATION_SEARCH",
              "ACCOMMODATION_DETAIL",
              "LIVE_PRICE",
              "AVAILABILITY",
              "RESTAURANT_SEARCH",
              "RESTAURANT_DETAIL",
              "ACTIVITY_SEARCH",
              "ACTIVITY_DETAIL",
              "FLIGHT_SEARCH",
              "CAR_RENTAL_SEARCH",
              "REVIEWS",
              "PHOTOS",
              "BOOKING_REDIRECT",
              "BOOKING_API",
              "AFFILIATE_LINK",
              "WEBHOOK",
              "CONVERSION_REPORTING"
            ]
          },
          "requirement": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "REQUIRED",
              "NOT_REQUIRED",
              "CONDITIONAL"
            ]
          },
          "displayText": {
            "type": "string"
          },
          "logoRequired": {
            "type": "boolean"
          },
          "linkUrl": {
            "type": "string"
          },
          "placementNotes": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "requirement"
        ]
      },
...
      "UpdateProviderLicenseDto": {
        "type": "object",
        "properties": {
          "datasetOrProduct": {
            "type": "string"
          },
          "licenseType": {
            "type": "string"
          },
          "termsUrl": {
            "type": "string"
          },
          "privacyUrl": {
            "type": "string"
          },
          "developerTermsUrl": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string"
          },
          "effectiveUntil": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "SetProviderLicenseRightsDto": {
        "type": "object",
        "properties": {
          "rightsDisplay": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "rightsCache": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "rightsStore": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "rightsModify": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "rightsRedistribute": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "rightsCommercialUse": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "conditionalNotes": {
            "type": "string",
            "description": "Required when any right above is CONDITIONAL."
          },
          "attributionRequirement": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "REQUIRED",
              "NOT_REQUIRED",
              "CONDITIONAL"
            ]
          }
        },
        "required": [
          "rightsDisplay",
          "rightsCache",
          "rightsStore",
          "rightsModify",
          "rightsRedistribute",
          "rightsCommercialUse",
          "attributionRequirement"
        ]
      },
      "SetProviderLicenseStatusDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProviderPolicyEvidenceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "accessedAt": {
            "type": "string"
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "OFFICIAL_TERMS",
              "OFFICIAL_DEVELOPER_DOC",
              "OFFICIAL_PARTNER_DOC",
              "OFFICIAL_PRICING_DOC",
              "CONTRACT",
              "LEGAL_REVIEW",
              "OTHER"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "sourceUrl",
          "accessedAt",
          "sourceType"
        ]
      },
      "CreateProviderDataPolicyDto": {
        "type": "object",
        "properties": {
          "capability": {
            "type": "string",
            "enum": [
              "PLACE_SEARCH",
              "PLACE_DETAIL",
              "NEARBY_SEARCH",
              "TEXT_SEARCH",
              "ACCOMMODATION_SEARCH",
              "ACCOMMODATION_DETAIL",
              "LIVE_PRICE",
              "AVAILABILITY",
              "RESTAURANT_SEARCH",
              "RESTAURANT_DETAIL",
              "ACTIVITY_SEARCH",
              "ACTIVITY_DETAIL",
              "FLIGHT_SEARCH",
              "CAR_RENTAL_SEARCH",
              "REVIEWS",
              "PHOTOS",
              "BOOKING_REDIRECT",
              "BOOKING_API",
              "AFFILIATE_LINK",
              "WEBHOOK",
              "CONVERSION_REPORTING"
            ]
          },
          "cacheAllowed": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "maxCacheSeconds": {
            "type": "number"
          },
          "storeIdentityAllowed": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "storeContentAllowed": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "refreshRequiredAfterSeconds": {
            "type": "number"
          },
          "deleteAfterSeconds": {
            "type": "number"
          },
          "persistentIdentifierAllowed": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "ALLOWED",
              "PROHIBITED",
              "CONDITIONAL"
            ]
          },
          "policySourceUrl": {
            "type": "string"
          }
        },
        "required": [
          "capability",
          "cacheAllowed",
          "storeIdentityAllowed",
          "storeContentAllowed",
          "persistentIdentifierAllowed"
        ]
      }
    }
  }
}